requirements.txt
asgiref==3.8.1
attrs==23.2.0
bcrypt==4.2.0
cffi==1.16.0
cryptography==43.0.0
Django==4.2.14
django-cors-headers==4.4.0
django-extensions==3.2.3
django-stronghold==0.4.0
django-webpack-loader==3.1.0
djangorestframework==3.15.2
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mysqlclient==2.2.4
numpy==2.0.0
pandas==2.2.2
parameterized==0.9.0
paramiko==3.4.0
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.5.0
sshtunnel==0.4.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
